+++ Sat.Recruitment.Api/Business/DTOs/UserDTO.cs
@@ -1,12 +1,12 @@
-﻿namespace Sat.Recruitment.Api.Entities
+﻿namespace Sat.Recruitment.Api.Business.DTOs
 {
-    public class User
+    public class UserDTO
     {
         public string Name { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
-        public UserType UserType { get; set; }
+        public string UserType { get; set; }
         public decimal Money { get; set; }
     }
 }
