+++ Sat.Recruitment.Api/Business/Entities/User.cs
@@ -1,12 +1,12 @@
-﻿namespace Sat.Recruitment.Api.DTOs
+﻿namespace Sat.Recruitment.Api.Business.Entities
 {
-    public class UserDTO
+    public class User
     {
         public string Name { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
-        public string UserType { get; set; }
+        public UserType UserType { get; set; }
         public decimal Money { get; set; }
     }
 }
