+++ Sat.Recruitment.Api/Business/Strategies/IGiftStrategy.cs
@@ -1,0 +1,9 @@
+﻿using Sat.Recruitment.Api.Business.Entities;
+
+namespace Sat.Recruitment.Api.Business.Strategies
+{
+    public interface IGiftStrategy
+    {
+        void ApplyGift(User user);
+    }
+}
