+++ Sat.Recruitment.Api/Business/Strategies/NormalGiftStrategy.cs
@@ -1,6 +1,6 @@
-﻿using Sat.Recruitment.Api.Entities;
+﻿using Sat.Recruitment.Api.Business.Entities;
 
-namespace Sat.Recruitment.Api.Strategies
+namespace Sat.Recruitment.Api.Business.Strategies
 {
     public class NormalGiftStrategy : BaseStrategy, IGiftStrategy
     {
