--- Sat.Recruitment.Api/Files/Users - Copy.txt
@@ -1,3 +1,0 @@
-Juan,Juan@marmol.com,+5491154762312,Peru 2464,Normal,1234
-Franco,Franco.Perez@gmail.com,+534645213542,Alvear y Colombres,Premium,112234
-Agustina,Agustina@gmail.com,+534645213542,Garay y Otra Calle,SuperUser,112234
