--- Sat.Recruitment.Api/Strategies/IGiftStrategy.cs
@@ -1,9 +1,0 @@
-﻿using Sat.Recruitment.Api.Entities;
-
-namespace Sat.Recruitment.Api.Strategies
-{
-    public interface IGiftStrategy
-    {
-        void ApplyGift(User user);
-    }
-}
